Optimize sequential reads in SearchableSnapshotIndexInput
Today `SearchableSnapshotIndexInput` translates each `readBytesInternal` call
to one or more calls to `readBlob` on the underlying repository. We make a lot
of small `readBytesInternal` calls since they are used to fill a small
in-memory buffer. Calls to `readBlob` are expensive: blob storage providers
like AWS S3 charge money per API call.

A common usage pattern is to take a brand-new `IndexInput`, seek to a
particular location, and then sequentially read a substantial amount of data
and stream it to disk.

This commit optimizes the implementation for that specific usage pattern.
Rather than calling `readBlob` each time the internal buffer needs filling we
instead request a (potentially much larger) range of the blob and consume the
response bit-by-bit as needed by a sequentially-reading client.

+++ server/src/main/java/org/elasticsearch/common/blobstore/BlobContainer.java
@@ -61,7 +61,7 @@ public interface BlobContainer {
      * @throws NoSuchFileException if the blob does not exist
      * @throws IOException         if the blob can not be read.
      */
-    default InputStream readBlob(final String blobName, final long position, final int length) throws IOException {
+    default InputStream readBlob(final String blobName, final long position, final long length) throws IOException {
         throw new UnsupportedOperationException(); // NORELEASE
     }
